requirements.txt
pandas
numpy
scikit-learn
geopy
streamlit
folium
matplotlib
seaborn
flask
python-dotenv
